2008-07-22  Aron Helser  <helser@cs.unc.edu>
	* nmb_CalculatedPlane.h (nmb_CalculatedPlane): make default
	constructor unimplemented again, so it can't be used.
	* nmb_FlattenedPlane.h (nmb_FlattenedPlane): remove 'virtual'
	keyword from 'public' inheritance, because it was used incorrectly
	and confused VS 2005.
	* nmb_*Plane.h : same
	* readAsylumFile.C : fix some compile warnings.

+++ src/lib/nmBase/readAsylumFile.C
@@ -399,7 +399,7 @@ int BCGrid::readAsylumFile(FILE* fr, const char *filename)
    if (fseek(fr, startFilePos, SEEK_SET) != 0) return -1;
    
    numBytesRead = fread(buffer, 1, binHeaderSize+waveHeaderSize, fr);
-   if (numBytesRead < binHeaderSize+waveHeaderSize || ferror(fr)) {
+   if ((int)numBytesRead < binHeaderSize+waveHeaderSize || ferror(fr)) {
       printf("Error occurred while reading the file headers.\n");
       return -1;
    }
@@ -556,7 +556,7 @@ int BCGrid::readAsylumFile(FILE* fr, const char *filename)
       memset(&(wave_name[0]), '\0', 32);
       
       numBytesRead = fread(notes_str, 1, num_bytes_left, fr);
-      if (numBytesRead < num_bytes_left || ferror(fr)) {
+      if ((long)numBytesRead < num_bytes_left || ferror(fr)) {
          delete [] notes_str;
          return -1;
       }
